Clean baseline: restore 31/31 blargg, 75/75 mooneye; strict -Werror build + ASan target
- Restore gb.c from 9b98849 (last known good for both test suites)
- Add strict compiler flags: -Wall -Wextra -Wconversion -Wsign-conversion
  -Wshadow -Wundef -Werror to default build target
- Add 'asan' Makefile target with -fsanitize=undefined,address -> gb_asan
- Fix all warnings in gb.c and bits.h (cast additions only, no logic changes):
  - Remove unused gpu_get_mode_for_lock function and timer_clocks array
  - Add (byte)/(word) casts throughout for implicit int conversion warnings
  - Fix PC += (int8_t)offset -> PC = (word)(PC + (int8_t)offset)
  - Fix cpu_write(--SP, ...) push patterns with (byte) casts
  - Fix AF/HL/DE/BC register getters with (word) outer cast
  - Fix RL/Rlc arg <<= 1 -> arg = (byte)(arg << 1)
  - Fix setFlag ternary, Swap, Sla, Sub return, oam_read_word
  - Fix Rr unused isA param, cart_load size_t casts
  - Fix INTERRUPT_PRIORITY char->byte cast at usage site
- Both make and make asan build with zero warnings/errors

+++ bits.h
@@ -1,4 +1,4 @@
-#define bit_check(source, bit) ((source & (1 << bit)) != 0)
-#define bit_set(source, bit) (source | (1 << bit))
-#define bit_clear(source, bit) (source & ~(1 << bit))
-#define bit_def(source, bit, doit) (doit ? bit_set(source, bit) : bit_clear(source, bit) )
+#define bit_check(source, bit) (((source) & (1u << (unsigned)(bit))) != 0u)
+#define bit_set(source, bit) ((byte)((source) | (1u << (unsigned)(bit))))
+#define bit_clear(source, bit) ((byte)((source) & ~(1u << (unsigned)(bit))))
+#define bit_def(source, bit, doit) ((doit) ? bit_set((source), (bit)) : bit_clear((source), (bit)))
